% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseConnect.R
\name{DatabaseConnector}
\alias{DatabaseConnector}
\title{DatabaseConnector}
\description{
R6 Class for connecting and interacting with various databases (MySQL,Azure SQL).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{con}}{Database connection object (NULL if disconnected)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseConnector-new}{\code{DatabaseConnector$new()}}
\item \href{#method-DatabaseConnector-connect_db}{\code{DatabaseConnector$connect_db()}}
\item \href{#method-DatabaseConnector-fetch_table_as_dataframe}{\code{DatabaseConnector$fetch_table_as_dataframe()}}
\item \href{#method-DatabaseConnector-add_record}{\code{DatabaseConnector$add_record()}}
\item \href{#method-DatabaseConnector-remove_record}{\code{DatabaseConnector$remove_record()}}
\item \href{#method-DatabaseConnector-disconnect}{\code{DatabaseConnector$disconnect()}}
\item \href{#method-DatabaseConnector-clone}{\code{DatabaseConnector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-new}{}}}
\subsection{Method \code{new()}}{
Initialize the connector (no active connection)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-connect_db"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-connect_db}{}}}
\subsection{Method \code{connect_db()}}{
Connect to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$connect_db(
  db_type,
  host,
  port,
  dbname,
  user,
  password,
  in_drv = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{db_type}}{Type of database: "mysql", "azure", "PostGres"}

\item{\code{host}}{Host address}

\item{\code{port}}{Port number (optional)}

\item{\code{dbname}}{Database name}

\item{\code{user}}{Username}

\item{\code{password}}{Password}

\item{\code{in_drv}}{Input Driver File (for Azure only)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE if connection successful, FALSE otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-fetch_table_as_dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-fetch_table_as_dataframe}{}}}
\subsection{Method \code{fetch_table_as_dataframe()}}{
Fetch a table as a DataFrame (tibble)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$fetch_table_as_dataframe(table_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Name of the table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble with data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-add_record"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-add_record}{}}}
\subsection{Method \code{add_record()}}{
Add a record to a table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$add_record(table_name, data_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Name of the table}

\item{\code{data_list}}{Named list of column = value pairs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-remove_record"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-remove_record}{}}}
\subsection{Method \code{remove_record()}}{
Remove a record from a table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$remove_record(table_name, col, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table_name}}{Name of the table}

\item{\code{col}}{Column name for condition}

\item{\code{value}}{Value to match}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
