% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{describe_overview.character}
\alias{describe_overview.character}
\title{Describe overview pour des character}
\usage{
\method{describe_overview}{character}(char)
}
\arguments{
\item{char}{un vecteur character}
}
\description{
Describe overview pour des character
}
