% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{describe_overview.default}
\alias{describe_overview.default}
\title{Méthode par défaut}
\usage{
\method{describe_overview}{default}(x)
}
\arguments{
\item{x}{N'importe quel objet ne rentrant pas dans les class implémentées}
}
\description{
Méthode par défaut
}
