% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{describe_overview.factor}
\alias{describe_overview.factor}
\title{Describe overview pour des facteurs}
\usage{
\method{describe_overview}{factor}(fact)
}
\arguments{
\item{fact}{un facteur}
}
\description{
Describe overview pour des facteurs
}
