% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{describe_overview.integer}
\alias{describe_overview.integer}
\title{Describe overview pour des vecteur d'integer}
\usage{
\method{describe_overview}{integer}(vect)
}
\arguments{
\item{vect}{un vecteur d'integer}
}
\description{
Describe overview pour des vecteur d'integer
}
