% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{describe_overview.numeric}
\alias{describe_overview.numeric}
\title{Describe overview pour des vecteur de numeric/double}
\usage{
\method{describe_overview}{numeric}(vect)
}
\arguments{
\item{vect}{un vecteur de numeric/double}
}
\description{
Describe overview pour des vecteur de numeric/double
}
