% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{summarize_salary}
\alias{summarize_salary}
\title{summarize_salary - mean/median/IQR}
\usage{
summarize_salary(
  x,
  applyto = "salary_in_usd",
  method = c("median", "iqr", "mean"),
  by = "work_year",
  na.rm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{applyto}{nom de colonne à laquelle appliquer la moyenne/medianne...}

\item{method}{moyenne/median/iqr.}

\item{by}{nom des colonnes à grouper pour appliquer.}

\item{na.rm}{TRUE, enlève les valeurs manquantes}

\item{verbose}{afficher un résumé des opérations.}
}
\value{
Le résultat par méthode groupé.
}
\description{
summarize_salary - mean/median/IQR
}
