% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_testing.R
\name{variance_homogeneity_test}
\alias{variance_homogeneity_test}
\title{variance_homogeneity_test - levene/bartlett}
\usage{
variance_homogeneity_test(
  data,
  y,
  group,
  method = c("levene", "bartlett"),
  na.rm = TRUE
)
}
\arguments{
\item{data}{dataset}

\item{y}{colonne à appliquer.}

\item{group}{nom de colonne à grouper pour appliquer.}

\item{method}{type de test}

\item{na.rm}{TRUE, enlève les valeurs manquantes}
}
\value{
Le résultat par méthode groupé.
}
\description{
variance_homogeneity_test - levene/bartlett
}
