% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{DataFrame}
\alias{DataFrame}
\title{DataFrame Class}
\description{
A custom R6 class that provides an interface for handling tabular data using tibbles.
It includes methods for initializing data, retrieving subsets, and performing basic table operations.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataFrame-new}{\code{DataFrame$new()}}
\item \href{#method-DataFrame-show_dp}{\code{DataFrame$show_dp()}}
\item \href{#method-DataFrame-get_data}{\code{DataFrame$get_data()}}
\item \href{#method-DataFrame-get_col}{\code{DataFrame$get_col()}}
\item \href{#method-DataFrame-get_size}{\code{DataFrame$get_size()}}
\item \href{#method-DataFrame-select}{\code{DataFrame$select()}}
\item \href{#method-DataFrame-test}{\code{DataFrame$test()}}
\item \href{#method-DataFrame-clone}{\code{DataFrame$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-new}{}}}
\subsection{Method \code{new()}}{
Initializes the DataFrame with named vectors as columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named vectors that form the columns of the DataFrame.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-show_dp"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-show_dp}{}}}
\subsection{Method \code{show_dp()}}{
Displays the stored tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$show_dp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-get_data}{}}}
\subsection{Method \code{get_data()}}{
Retrieves the entire stored tibble.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble containing all the stored data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-get_col"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-get_col}{}}}
\subsection{Method \code{get_col()}}{
Retrieves one or more columns by name.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$get_col(colName)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{colName}}{A string or a vector of strings specifying the column(s) to retrieve.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble containing the selected columns.
throws an Error if \code{colName} is not a character or if a specified column does not exist.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-get_size"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-get_size}{}}}
\subsection{Method \code{get_size()}}{
Retrieves the dimensions of the table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$get_size()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Also prints a message with the table size.
}

\subsection{Returns}{
A numeric vector containing the number of rows and columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-select"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-select}{}}}
\subsection{Method \code{select()}}{
Selects specific rows and columns from the DataFrame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$select(id_row = NULL, id_col = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_row}}{A numeric vector of row indices to select. Defaults to all rows if NULL.}

\item{\code{id_col}}{A numeric vector of column indices to select. Defaults to all columns if NULL.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A tibble subset of \code{private$data}.
throws an  Error if row or column indices exceed table dimensions.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-test"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-test}{}}}
\subsection{Method \code{test()}}{
For testing purpose
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$test()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataFrame-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataFrame-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataFrame$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
